Modify iked to once again send a fully detailed checkpoint vendor id. This was pulled in a previous commit to simplify the code during trouble shooting.
git-svn-id: svn://svn.shrew.net/ike/head@317 1bb156eb-19af-db11-b919-0014c23ffdf2

+++ source/iked/ike.cpp
@@ -492,7 +492,7 @@ long _IKED::packet_ike_decrypt( IDB_PH1 * sa, PACKET_IKE & packet, BDATA * iv )
 			 diff );
 
 		header = ( IKE_HEADER * ) packet.buff();
-		header->length = htonl( size );
+		header->length = htonl( ( uint32_t ) size );
 	}
 
 	//
@@ -561,7 +561,7 @@ long _IKED::packet_ike_encrypt( IDB_PH1 * sa, PACKET_IKE & packet, BDATA * iv )
 	//
 
 	IKE_HEADER * header = ( IKE_HEADER * ) packet.buff();
-	header->length = htonl( size );
+	header->length = htonl( ( uint32_t ) size );
 
 	//
 	// init cipher key and iv and
